chore: 드래그 훅 수정
